+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [0.15.7] - Unreleased
+### Added
+- `manifest.json` file to install the CMS as a PWA.
+
 ### Fixed
 - Search placeholder translation.